+++ src/actions/synchronization/taskunifier/TaskUnifierAccountInfoActions.js
@@ -1,4 +1,5 @@
 import { sendRequest } from 'actions/RequestActions';
+import { updateSettings } from 'actions/SettingActions';
 import { setSynchronizationData } from 'actions/SynchronizationActions';
 import { getConfig } from 'config/Config';
 import { getSettings } from 'selectors/SettingSelectors';
@@ -18,6 +19,13 @@ export function getTaskUnifierAccountInfo() {
             settings);
 
         console.debug('getTaskUnifierAccountInfo', result);
+
+        await dispatch(updateSettings({
+            taskunifier: {
+                ...settings.taskunifier,
+                accountInfo: result.data
+            }
+        }));
 
         await dispatch(setSynchronizationData('taskunifier', {
             accountInfo: result.data